+++ src/main/groovy/uk/ac/ox/softeng/maurodatamapper/plugins/database/SamplingStrategy.groovy
@@ -22,7 +22,7 @@ class SamplingStrategy {
     boolean enabled
     Integer threshold
     BigDecimal percentage
-    Integer approxCount
+    Long approxCount
 
     SamplingStrategy() {
         this.enabled = false
